lambda/requirements.txt
Pillow>=9.5.0,<11.0.0